third_party/blink/web_tests/media/encrypted-media/encrypted-media-getstatusforpolicy.html

<!DOCTYPE html>
<html>
    <head>
        <title>MediaKeys.getStatusForPolicy</title>
        <script src="encrypted-media-utils.js"></script>
        <script src="../../resources/testharness.js"></script>
        <script src="../../resources/testharnessreport.js"></script>
    </head>
    <body>
        <script>
            promise_test(function(test)
            {
                var hdcpVersionNone = '';
                var hdcpVersion2_0 = '2.0';
                var mediaKeys;

                // Calls getStatusForPolicy() with |hdcpVersion| and expects it
                // to resolve.
                function getStatusForHdcpPolicy(hdcpVersion, expectedError)
                {
                    return mediaKeys.getStatusForPolicy(hdcpVersion).then(function(MediaKeyStatus) {
                        // getStatusForPolicy() is supported by Clear Key key
                        // system and the promise should always be resolved.
                        return Promise.resolve();
                    }, function(error) {
                        if (expectedError === undefined) {
                            assert_equals(error.name, 'NotSupportedError');
                        } else {
                            assert_equals(error, expectedError);
                        }
                        return Promise.resolve('Error encountered as expected.');
                    });
                };

                return navigator.requestMediaKeySystemAccess('org.w3.clearkey', getSimpleConfiguration()).then(function(access) {
                    return access.createMediaKeys();
                }).then(function(result) {
                    mediaKeys = result;
                    return getStatusForHdcpPolicy({});
                }).then(function() {
                    return getStatusForHdcpPolicy({minHdcpVersion: hdcpVersionNone});
                }).then(function() {
                    return getStatusForHdcpPolicy({minHdcpVersion: hdcpVersion2_0});
                }).then(function() {
                    return getStatusForHdcpPolicy({foo: 'bar'});
                }).then(function() {
                    return getStatusForHdcpPolicy(
                        { get minHdcpVersion() { return hdcpVersion2_0; } });
                }).then(function() {
                    return getStatusForHdcpPolicy(
                        { get minHdcpVersion() { throw 'throwing getter'; } },
                        'throwing getter');
                });
            }, 'MediaKeys.getStatusForPolicy');
        </script>
    </body>
</html>
